chore: deduplicate code
